# docs/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ditto Quant API",
    "description": "量化投资系统API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/universes": {
      "get": {
        "tags": [
          "universes"
        ],
        "summary": "List Universes",
        "description": "列出所有 Universe.",
        "operationId": "list_universes_api_v1_universes_get",
        "parameters": [
          {
            "name": "universe_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "类型过滤",
              "title": "Universe Type"
            },
            "description": "类型过滤"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_UniverseResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "universes"
        ],
        "summary": "Create Universe",
        "description": "创建自定义 Universe.",
        "operationId": "create_universe_api_v1_universes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUniverseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_UniverseResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/universes/{universe_id}": {
      "get": {
        "tags": [
          "universes"
        ],
        "summary": "Get Universe",
        "description": "获取 Universe 详情.",
        "operationId": "get_universe_api_v1_universes__universe_id__get",
        "parameters": [
          {
            "name": "universe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Universe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_UniverseResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "universes"
        ],
        "summary": "Update Universe",
        "description": "更新自定义 Universe.",
        "operationId": "update_universe_api_v1_universes__universe_id__put",
        "parameters": [
          {
            "name": "universe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Universe Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUniverseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_UniverseResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "universes"
        ],
        "summary": "Delete Universe",
        "description": "删除自定义 Universe（预设不可删）.",
        "operationId": "delete_universe_api_v1_universes__universe_id__delete",
        "parameters": [
          {
            "name": "universe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Universe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_bool_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/universes/{universe_id}/members": {
      "get": {
        "tags": [
          "universes"
        ],
        "summary": "Get Members",
        "description": "获取 Universe 成分股.",
        "operationId": "get_members_api_v1_universes__universe_id__members_get",
        "parameters": [
          {
            "name": "universe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Universe Id"
            }
          },
          {
            "name": "asof",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "PIT 日期",
              "title": "Asof"
            },
            "description": "PIT 日期"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_MemberResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_UniverseResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UniverseResponse",
            "description": "响应数据"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "分页信息(可选)"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[UniverseResponse]"
      },
      "APIResponse_bool_": {
        "properties": {
          "data": {
            "type": "boolean",
            "title": "Data",
            "description": "响应数据"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "分页信息(可选)"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[bool]"
      },
...
      "APIResponse_list_MemberResponse__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            },
            "type": "array",
            "title": "Data",
            "description": "响应数据"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "分页信息(可选)"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[list[MemberResponse]]"
      },
...
      "APIResponse_list_UniverseResponse__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/UniverseResponse"
            },
            "type": "array",
            "title": "Data",
            "description": "响应数据"
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "分页信息(可选)"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[list[UniverseResponse]]"
      },
...
      "CreateUniverseRequest": {
        "properties": {
          "universe_id": {
            "type": "string",
            "minLength": 1,
            "title": "Universe Id",
            "description": "Universe ID"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "名称"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "描述"
          }
        },
        "type": "object",
        "required": [
          "universe_id",
          "name"
        ],
        "title": "CreateUniverseRequest",
        "description": "创建 Universe 请求."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MemberResponse": {
        "properties": {
          "instrument_id": {
            "type": "integer",
            "title": "Instrument Id"
          }
        },
        "type": "object",
        "required": [
          "instrument_id"
        ],
        "title": "MemberResponse",
        "description": "Universe 成分股响应."
      },
      "PaginationResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "总记录数"
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "description": "每页数量"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "当前偏移量"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "判断是否还有更多数据.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PaginationResponse",
        "description": "分页响应模型.\n\nAttributes:\n    total: 总记录数.\n    limit: 每页数量.\n    offset: 当前偏移量.\n    has_more: 是否有更多数据(计算字段)."
      },
...
      "UniverseResponse": {
        "properties": {
          "universe_id": {
            "type": "string",
            "title": "Universe Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "universe_type": {
            "type": "string",
            "title": "Universe Type",
            "default": "custom"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          }
        },
        "type": "object",
        "required": [
          "universe_id",
          "name"
        ],
        "title": "UniverseResponse",
        "description": "Universe 响应."
      },
...
      "UpdateUniverseRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name",
            "description": "名称"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "描述"
          },
          "members": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Members",
            "description": "成分列表"
          },
          "effective_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date",
            "description": "生效日期"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "UpdateUniverseRequest",
        "description": "更新 Universe 请求."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}